Moved tinker13 models into their own module
