+++ halotools/empirical_models/occupation_models/__init__.py
@@ -1,3 +1,4 @@
 from .hod_components import *
 from .zheng07_components import *
 from .leauthaud11_components import *
+from .tinker13_components import *
