+++ halotools/empirical_models/occupation_models/leauthaud11_components.py
@@ -21,10 +21,11 @@ from astropy.extern import six
 from abc import ABCMeta, abstractmethod, abstractproperty
 import warnings
 
+from .hod_components import OccupationComponent 
+
 from .. import model_defaults, model_helpers, smhm_components
 from ..assembias_models import HeavisideAssembias
 from ..model_helpers import bounds_enforcing_decorator_factory
-from .hod_components import OccupationComponent 
 
 from ...utils.array_utils import custom_len
 from ...utils.table_utils import compute_conditional_percentiles
